+++ __tests__/service/play.test.js
@@ -1,0 +1,6 @@
+//  [ ] 테스트 코드 수정 필오
+describe('play(cars, totalRound, moveStrategies) 테스트', () => {
+    it('test description', () => {
+        expect(1).toBe(1);
+    });
+});
