+++ src/Models/service/Cars/createCars.js
@@ -1,0 +1,5 @@
+import Car from '../../entities/Car/Car.js';
+
+export default function createCars(carNames) {
+    return carNames.map((carName) => Car.of(carName));
+}
