--- src/Models/service/InvalidCallOrderError.js
@@ -1,12 +1,0 @@
-import RuntimeError from '../../RuntimeError.js';
-
-export default class InvalidCallOrderError extends RuntimeError {
-    #type;
-    #message;
-
-    constructor(message) {
-        super(message);
-        this.#type = this.constructor.name;
-        this.#message = message;
-    }
-}
