+++ src/Models/service/Rounds/getSnapshots.js
@@ -1,0 +1,3 @@
+export default function getRoundSnapshots(rounds) {
+    return rounds.map((round) => round.snapshot);
+}
