+++ src/Models/service/ValidationError.js
@@ -1,0 +1,10 @@
+import RuntimeError from '../../RuntimeError.js';
+
+// 사용자 입력값 검증 오류
+export default class ValidationError extends RuntimeError {
+    static #type = 'SERVICE';
+
+    constructor(message) {
+        super(ValidationError.#type, message);
+    }
+}
