--- src/Models/service/errors.js
@@ -1,93 +1,0 @@
-import RuntimeError from '../../RuntimeError.js';
-
-// 사용자 입력값 검증 오류
-class ValidationError extends RuntimeError {
-    static #type = 'SERVICE';
-
-    constructor(message) {
-        super(ValidationError.#type, message);
-    }
-}
-
-export class CarNamesNotStringError extends ValidationError {
-    static #MESSAGE = '사용자는 자동차 이름들을 문자열 형태로 입력해야합니다.';
-
-    constructor() {
-        super(CarNamesNotStringError.#MESSAGE);
-    }
-}
-
-export class CarNamesEmptyStringError extends ValidationError {
-    static #MESSAGE = '사용자는 빈 값으로 자동차 이름들을 입력해서는 안됩니다.';
-
-    constructor() {
-        super(CarNamesEmptyStringError.#MESSAGE);
-    }
-}
-
-export class CarNamesDuplicatedError extends ValidationError {
-    static #MESSAGE =
-        '사용자는 자동차 이름들로 중복되지 않은 자동차 이름을 입력해야합니다.';
-
-    constructor() {
-        super(CarNamesDuplicatedError.#MESSAGE);
-    }
-}
-
-export class TotalRoundEmptyError extends ValidationError {
-    static #MESSAGE = '사용자는 시도 횟수로 빈 값을 입력해서는 안됩니다.';
-
-    constructor() {
-        super(TotalRoundEmptyError.#MESSAGE);
-    }
-}
-
-export class TotalRoundNotNumberError extends ValidationError {
-    static #MESSAGE = '사용자는 시도 횟수로 숫자 형태를 입력해야합니다.';
-
-    constructor() {
-        super(TotalRoundNotNumberError.#MESSAGE);
-    }
-}
-
-export class TotalRoundNotIntegerError extends ValidationError {
-    static #MESSAGE = '사용자는 시도 횟수로 정수를 입력해야합니다.';
-
-    constructor() {
-        super(TotalRoundNotIntegerError.#MESSAGE);
-    }
-}
-
-export class TotalRoundOutOfRangeError extends ValidationError {
-    static #MESSAGE =
-        '사용자는 시도 횟수로는 1 이상 10 이하의 숫자를 입력해야합니다.';
-
-    constructor() {
-        super(TotalRoundOutOfRangeError.#MESSAGE);
-    }
-}
-
-export class MoveStrategiesNotArrayError extends RuntimeError {
-    static #MESSAGE = '이동 전략들은 배열 형태로 입력해주세요.';
-
-    constructor() {
-        super(MoveStrategiesNotArrayError.#MESSAGE);
-    }
-}
-
-export class MoveStrategiesElementNotMoveStrategyError extends RuntimeError {
-    static #MESSAGE =
-        '이동 전략들은 MoveStrategy의 인스턴스로 구성되어야합니다.';
-
-    constructor() {
-        super(MoveStrategiesElementNotMoveStrategyError.#MESSAGE);
-    }
-}
-
-export class MovesStrategiesLengthError extends RuntimeError {
-    static #MESSAGE = '이동 전략의 개수는 자동차의 수와 같아야합니다.';
-
-    constructor() {
-        super(MovesStrategiesLengthError.#MESSAGE);
-    }
-}
