+++ src/Models/service/play.js
@@ -1,0 +1,5 @@
+export default function play(cars, rounds, moveStrategies) {
+    rounds.map((round) => {
+        cars = round.run(cars, moveStrategies);
+    });
+}
